Configured celerybeat with celery-redbeat
