+++ BackEnd/news-service/app/celery.py
@@ -15,3 +15,10 @@ celery = Celery(
     include=["app.tasks"]
 )
 
+# Configure periodic task for Celery Beat
+celery.conf.beat_scheduler = "redbeat.RedBeatScheduler"
+celery.conf.redbeat_lock_key = None  # Disable locking issues
+
+
+
+celery.conf.timezone = "UTC"
